Question6

+++ CalculateSimpleInterest.java
@@ -5,7 +5,7 @@ Scanner kb=new Scanner(System.in);
 int p=kb.nextInt();
 int r=kb.nextInt();
 int t=kb.nextInt();
-double SI=(p*t*r)/100;
+int SI=(p*t*r)/100;
 
 
 System.out.println("SI is "+ SI);
